feat(build): VIXEN fully builds on WSL/GCC (GLFW windowing port + MSVC->GCC portability)
GLFW replaces raw Win32 windowing everywhere (WindowNode/SwapChainNode/InputNode/VulkanSwapChain/
CapabilityGraph/app-loop -> glfwCreateWindow / glfwCreateWindowSurface / glfwGetRequiredInstance-
Extensions / glfwGetKey / glfwPollEvents; HWND+HINSTANCE slots -> one GLFWwindow* WINDOW slot; zero
#ifdef _WIN32 in the windowing path). Plus MSVC->GCC portability fixes across non-windowing
subsystems the port then reached (ResourceManagement const-default-arg, Core trackers, missing
includes, const-on-struct, two-phase-lookup), and CMake wiring (link glfw, Gaia single-header
toggle, Vulkan loader-path resolve).

Result: cmake --build is GREEN on WSL/GCC -- engine + both executables, zero manual setup. Windows/
MSVC build to be verified on merge (GLFW uses Win32 there; the portability fixes are MSVC-safe).

Found building VIXEN as the Undertow render engine.

+++ VIXEN/libraries/SVO/include/SVOLOD.h
@@ -86,7 +86,7 @@ struct LODParameters {
 
         LODParameters params;
         params.rayOrigSize = 0.0f;  // Pinhole camera: zero diameter at origin
-        params.rayDirSize = 2.0f * std::tanf(pixelAngle * 0.5f);  // Cone spread
+        params.rayDirSize = 2.0f * std::tan(pixelAngle * 0.5f);  // Cone spread
 
         return params;
     }
@@ -108,7 +108,7 @@ struct LODParameters {
         float nearPlane)
     {
         float pixelAngle = fovY / static_cast<float>(screenHeight);
-        float dirSize = 2.0f * std::tanf(pixelAngle * 0.5f);
+        float dirSize = 2.0f * std::tan(pixelAngle * 0.5f);
 
         LODParameters params;
         params.rayDirSize = dirSize;
@@ -169,7 +169,7 @@ struct LODParameters {
     [[nodiscard]] LODParameters withBias(float bias) const {
         LODParameters biased = *this;
         // Bias multiplies the cone spread - larger spread = earlier termination
-        float multiplier = std::powf(2.0f, bias);
+        float multiplier = std::pow(2.0f, bias);
         biased.rayDirSize *= multiplier;
         biased.rayOrigSize *= multiplier;
         return biased;
@@ -190,7 +190,7 @@ struct LODParameters {
     // ESVO: scale_exp2 = 2^(scale - 23) in [1,2] normalized space
     // World size = scale_exp2 * worldSize (since [1,2] maps to world bounds)
     constexpr int ESVO_MAX_SCALE = 22;
-    float normalizedSize = std::ldexpf(1.0f, esvoScale - ESVO_MAX_SCALE - 1);
+    float normalizedSize = std::ldexp(1.0f, esvoScale - ESVO_MAX_SCALE - 1);
     return normalizedSize * worldSize;
 }
